# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Template Simbi API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/loans": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Lista todos os empréstimos cadastrados na API",
                "description": "Lista todos os empréstimos cadastrados.",
                "operationId": "da1da231dd673e1cd157ba27ce4f70c9",
                "responses": {
                    "200": {
                        "description": "Loans list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Loan"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição com erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bad Request"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Proibido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Internal server error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Adiciona um novo empréstimo na API",
                "description": "Permite adicionar um novo empréstimo",
                "operationId": "bf75f8441c0b3e6818b9a5a48f2e512a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bookId"
                                ],
                                "properties": {
                                    "bookId": {
                                        "type": "string"
                                    },
                                    "loanDate": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "returnDate": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Loan Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição com erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bad Request"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Proibido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Internal server error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/authors/{id}/books": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Lista todos os livros cadastrados na API para um determinado autor",
                "description": "Lista todos os livros cadastrados por um determinado autor.",
                "operationId": "05e289382ad925246b99df38487e708f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do autor",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "2234f840-417e-4944-ac9b-e7e6eb06c590"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Books list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição com erro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Bad Request"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Proibido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Forbidden Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthorized Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Resource Not Found Error"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Internal server error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Author": {
                "title": "Author",
                "description": "Author entity",
                "properties": {
                    "id": {
                        "description": "ID of the author",
                        "type": "string",
                        "example": "1"
                    },
                    "name": {
                        "description": "Full name of the author",
                        "type": "string"
                    },
                    "createdAt": {
                        "description": "Timestamp when the author was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-29T04:25:50.000000Z"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the author was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-29T04:25:50.000000Z"
                    }
                },
                "type": "object"
            },
...
            "Book": {
                "title": "Book",
                "description": "Book entity",
                "properties": {
                    "id": {
                        "description": "ID of the book",
                        "type": "string",
                        "example": "1"
                    },
                    "title": {
                        "description": "Title of the book",
                        "type": "string",
                        "example": "Curso avançado de Java"
                    },
                    "publisher": {
                        "description": "Publisher of the book",
                        "type": "string",
                        "example": "Editora XPTO"
                    },
                    "authorId": {
                        "description": "ID of the author",
                        "type": "string",
                        "example": "1"
                    },
                    "author": {
                        "$ref": "#/components/schemas/Author"
                    },
                    "isAvailable": {
                        "description": "Availability status of the book",
                        "type": "boolean",
                        "example": true
                    },
                    "createdAt": {
                        "description": "Timestamp when the book was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-29T04:25:50.000000Z"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the book was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-29T04:25:50.000000Z"
                    }
                },
                "type": "object"
            },
            "Loan": {
                "title": "Loan",
                "description": "Loan entity",
                "properties": {
                    "id": {
                        "description": "ID of the loan",
                        "type": "string",
                        "example": "1"
                    },
                    "loanDate": {
                        "description": "Date when the loan was made",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-29T04:25:50.000000Z"
                    },
                    "returnDate": {
                        "description": "Date when the loan is expected to be returned",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-29T04:25:50.000000Z"
                    },
                    "status": {
                        "description": "Status of the loan",
                        "type": "string",
                        "enum": [
                            "created",
                            "active",
                            "finished",
                            "overdue"
                        ],
                        "example": "active"
                    },
                    "returnedAt": {
                        "description": "Date when the loan was returned",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-15T04:25:50.000000Z",
                        "nullable": true
                    },
                    "renewalCount": {
                        "description": "Number of times the loan has been renewed",
                        "type": "integer",
                        "example": 1
                    },
                    "lastRenewedAt": {
                        "description": "Date when the loan was last renewed",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-09-10T04:25:50.000000Z"
                    },
                    "book": {
                        "$ref": "#/components/schemas/Book"
                    },
                    "createdAt": {
                        "description": "Timestamp when the loan was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-29T04:25:50.000000Z"
                    },
                    "updatedAt": {
                        "description": "Timestamp when the loan was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-29T04:25:50.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}